Fix incorrect UQM_INT64 definition
Apparently UQM has never targeted 64-bit Windows, because it
assumed that `long` would always be 64-bits and thus defined
`UQM_INT64` incorrectly.  This uses an #ifdef to fix it.

types.h is filled with elaborate checks to make types that work on
all sorts of obsolete C versions and broken implementations, which
I would just as soon replace with <stdint.h>, but to minimize code
changes at this stage I'll leave that for later.