HACK: Remove HAVE_WINT_T check and say we always have it
This check does not seem to work reliably on all platforms, and I
couldn't find a check that does with some modest effort.  Because wint_t
is required by the C standard I'm not going to lose any sleep over failing
to accomodate systems that don't have it for some reason.