CMake: Replace uqm_lib_graphics with uqm_lib_sdl
`uqm_lib_graphics` was originally supposed to be a combination of
an SDL version and optionally OpenGL and some other things. But it
turns out SDL is needed for other things besides graphics, so it's
more useful to just have an alias for whichever SDL we're using.
OpenGL can be linked separately if needed.