CMake: Move uqm executable target up to sc2 directory
I had originally put `uqm` in sc2/src/ because that's where uqm.c
is, and all other project files are within sc2/src/.  But with the
executable named `uqm`, that conflicts with the `uqm` subfolder.
The original build system put the executable one folder up in sc2/,
so I'm moving it there.

In the process I'm adding yet another object library target, making
the build system still messier and more inconsistent.  I'll need to
revisit the design of all this later.