support for vita filesystem paths
new error to fix: "Fatal error: Could not find content"
paths seem to be accepted and work now, i think?

this filesystem stuff is kinda the worst and my implementation
is ridiculously hacky. this took many hours, this uio code is
not fun but i think i'm getting to the end of it now, let's
hope...

i will remove all the crazy debug printfs in a future commit
still not sure if i will need those for debugging next time