preliminary directional joystick support
ref issue #1
still WIP as there are a few bugs, also the option to enable the
directional joystick does not yet show, even though the config option is
there in the config file. so far only works in melee

majority of code is from the android port of the game

i think i'll leave this disabled by default because it's very imprecise,
i was expecting it to work a little bit better. this is because
sometimes the ship is slow and turns in the wrong direction you wanted
just because getting there is faster, so it has to assume that