full directional joystick implementation
closes #1

now works in hyperspace/truespace and on planets, when using the lander.
there is no longer the weird mechanic where the ship accelerates when
tapping the joystick. adjusted for minor stick drift, so that the ship
doesn't eternally spin if the stick is drifting a little and now we can
use the positive vertical axis of the second stick to use as a thrust
control (works pretty well)

i wanted to implement this as a setting in the setup menu, but i think
it's fine to leave it as it is. it is possible to switch between the two
controls without hassle so i think i'll just leave it as default. people
can always change the config option in the file which should work. maybe
i'll add the settings option if people want it in rev2